Сервис media-server запущен в инфрастуктуре docker compose
