+++ .gitignore
@@ -1,6 +1,8 @@
 command.txt
 docker-compose.override.yml
 
+hls/
+
 # Created by https://www.gitignore.io
 
 ### OSX ###
