+++ media_server/src/core/logger.py
@@ -1,6 +1,5 @@
 import structlog
 
-
 LOGGING = {
     "version": 1,
     "disable_existing_loggers": False,
