+++ media_server/src/entrypoint-fastapi.sh
