+++ party_manager_service/src/.env.fastapi
