+++ party_manager_service/src/Dockerfile
