+++ party_manager_service/src/api/v1/auth.py
