+++ party_manager_service/src/entrypoint-fastapi.sh
