+++ party_manager_service/src/integration/mongodb.py
