+++ party_manager_service/src/integration/rabbitmq.py
