+++ party_manager_service/src/integration/storages.py
