+++ party_manager_service/src/integration/websocket.py
