+++ party_manager_service/src/requirements.txt
